Move Main from App to EntryPoint
